requirements.txt
pdfplumber==0.11.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0

